Use gradle
