+++ build.gradle
@@ -1,0 +1,39 @@
+group 'com.elikill58'
+version '1.4'
+
+apply plugin: 'java'
+
+apply plugin: 'eclipse'
+apply plugin: 'idea'
+
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+
+repositories {
+    // spigot api
+    maven {
+        url 'https://hub.spigotmc.org/nexus/content/groups/public/'
+    }
+    // bungeecord stuff
+    maven {
+        url 'https://oss.sonatype.org/content/repositories/snapshots/'
+    }
+}
+
+dependencies {
+    compile 'org.spigotmc:spigot-api:1.11.2-R0.1-SNAPSHOT'
+}
+
+eclipse {
+    classpath {
+        downloadJavadoc = true
+        downloadSources = false
+    }
+}
+
+idea {
+    module {
+        downloadJavadoc = true
+        downloadSources = false
+    }
+}
