--- pasltime.jardesc
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="WINDOWS-1252" standalone="no"?>
-<jardesc>
-    <jar path="C:/Users/Mathis Soenen/Documents/Serveurs_locaux/Serveurs_minecraft/1.8.X/plugins/Ping.jar"/>
-    <options buildIfNeeded="true" compress="true" descriptionLocation="/Ping/pasltime.jardesc" exportErrors="true" exportWarnings="true" includeDirectoryEntries="false" overwrite="true" saveDescription="true" storeRefactorings="false" useSourceFolders="false"/>
-    <storedRefactorings deprecationInfo="true" structuralOnly="false"/>
-    <selectedProjects/>
-    <manifest generateManifest="true" manifestLocation="" manifestVersion="1.0" reuseManifest="false" saveManifest="false" usesManifest="true">
-        <sealing sealJar="false">
-            <packagesToSeal/>
-            <packagesToUnSeal/>
-        </sealing>
-    </manifest>
-    <selectedElements exportClassFiles="true" exportJavaFiles="false" exportOutputFolder="false">
-        <javaElement handleIdentifier="=Ping/src"/>
-        <file path="/Ping/.project"/>
-        <file path="/Ping/plugin.yml"/>
-        <file path="/Ping/.classpath"/>
-        <file path="/Ping/config.yml"/>
-        <file path="/Ping/pasltime.jardesc"/>
-    </selectedElements>
-</jardesc>
