--- plugin.yml
@@ -1,8 +1,0 @@
-name: EliPing
-main: com.elikill58.ping.Main
-version: 1.3
-author: Elikill58
-
-commands:
-   ping:
-    permission: ping.ping
