--- src/com/elikill58/ping/.gitignore
@@ -1,4 +1,0 @@
-/Main.class
-/PingCommand.class
-/Timer.class
-/Utils.class
